+++ src/app/core/code/code.service.ts
@@ -8,75 +8,21 @@ import { deepCopy } from '@delon/util/other';
 import type { NzSafeAny } from 'ng-zorro-antd/core/types';
 
 import angularJSON from './files/angular.json';
-import appModuleTS from './files/app.module';
-import delonABCModuleTS from './files/delon-abc.module';
-import delonChartModuleTS from './files/delon-chart.module';
-import environmentTS from './files/environment';
-import globalConfigTS from './files/global-config.module';
+import appConfigTS from './files/app.config';
 import mainTS from './files/main';
-import mainCliTS from './files/main-cli';
 import mockUser from './files/mock-user';
-import nzZorroAntdModuleTS from './files/ng-zorro-antd.module';
 import packageJSON from './files/package.json';
-import polyfillTS from './files/polyfill';
 import readme from './files/readme-cli';
-import sandboxConfigJSON from './files/sandbox.config.json';
+import sandboxConfigJSON from './files/sandbox';
 import startupServiceTS from './files/startup.service';
 import tsconfigJSON from './files/tsconfig.json';
+import yarnLock from './files/yarn.lock';
 import pkg from '../../../../package.json';
 import { AppService } from '../app.service';
 
 @Injectable({ providedIn: 'root' })
 export class CodeService {
   private document: Document;
-
-  // private get dependencies(): { [key: string]: string } {
-  //   const res: { [key: string]: string } = {};
-  //   [
-  //     '@angular/animations',
-  //     '@angular/compiler',
-  //     '@angular/common',
-  //     '@angular/core',
-  //     '@angular/forms',
-  //     '@angular/platform-browser',
-  //     '@angular/platform-browser-dynamic',
-  //     '@angular/router',
-  //     '@ant-design/icons-angular',
-  //     'core-js@3.8.3',
-  //     'rxjs',
-  //     'tslib',
-  //     'zone.js',
-  //     'date-fns',
-  //     `@angular/cdk@^${MAX_MAIN_VERSION}.x`,
-  //     'ng-zorro-antd',
-  //     '@delon/theme',
-  //     '@delon/abc',
-  //     '@delon/chart',
-  //     '@delon/acl',
-  //     '@delon/auth',
-  //     '@delon/cache',
-  //     '@delon/mock',
-  //     '@delon/form',
-  //     '@delon/util',
-  //     'ajv',
-  //     'ajv-formats'
-  //   ].forEach(key => {
-  //     const includeVersion = key.lastIndexOf(`@`);
-  //     if (includeVersion > 1) {
-  //       res[key.substring(0, includeVersion)] = key.substring(includeVersion + 1);
-  //       return;
-  //     }
-  //     const version = key.startsWith('@delon')
-  //       ? `~${pkg.version}`
-  //       : (
-  //           (pkg.dependencies || pkg.devDependencies) as {
-  //             [key: string]: string;
-  //           }
-  //         )[key];
-  //     res[key] = version || '*';
-  //   });
-  //   return res;
-  // }
 
   private get themePath(): string {
     return `node_modules/@delon/theme/${this.appSrv.theme}.css`;
@@ -96,6 +42,7 @@ export class CodeService {
     const res = packageJSON as Record<string, NzSafeAny>;
     [
       'ng-zorro-antd',
+      'ng-antd-color-picker',
       'date-fns',
       '@delon/theme',
       '@delon/abc',
@@ -134,7 +81,6 @@ export class CodeService {
     });
     // res.dependencies['core-js'] = `~3.8.3`;
     if (!includeCli) res;
-    console.log(res);
 
     return res;
   }
@@ -179,12 +125,21 @@ export class CodeService {
     };
   }
 
+  private attachStandalone(code: string): string {
+    // standalone: true,
+    if (code.includes(`standalone: true`)) return code;
+
+    return `${code.replace(`@Component({`, `@Component({\n  standalone: true,\n`)}`;
+  }
+
   openOnStackBlitz(title: string, appComponentCode: string): void {
+    appComponentCode = this.attachStandalone(appComponentCode);
     const res = this.parseCode(appComponentCode);
     const json = deepCopy(angularJSON);
     json.projects.demo.architect.build.options.styles.splice(0, 0, this.themePath);
     const packageJson = this.genPackage({ dependencies: [], devDependencies: [], includeCli: false });
-    packageJson.name = title;
+    packageJson.name = 'NG-ALAIN';
+    packageJson.description = title;
     sdk.openProject(
       {
         title: 'NG-ALAIN',
@@ -195,24 +150,30 @@ export class CodeService {
           ...(packageJson.devDependencies as Record<string, string>)
         },
         files: {
+          '.stackblitzrc': JSON.stringify(
+            {
+              installDependencies: true,
+              startCommand: 'yarn start',
+              env: {
+                ENABLE_CJS_IMPORTS: true
+              }
+            },
+            null,
+            2
+          ),
+          'yarn.lock': yarnLock,
           'angular.json': `${JSON.stringify(json, null, 2)}`,
           'tsconfig.json': `${JSON.stringify(tsconfigJSON, null, 2)}`,
           'package.json': `${JSON.stringify(packageJson, null, 2)}`,
-          'src/environments/environment.ts': environmentTS,
           'src/index.html': res.html,
-          'src/main.ts': mainTS,
-          'src/polyfills.ts': polyfillTS,
+          'src/main.ts': mainTS(res.componentName),
           'src/app/app.component.ts': appComponentCode,
-          'src/app/app.module.ts': appModuleTS(res.componentName),
-          'src/app/global-config.module.ts': globalConfigTS,
-          'src/app/ng-zorro-antd.module.ts': nzZorroAntdModuleTS,
-          'src/app/delon-abc.module.ts': delonABCModuleTS,
-          'src/app/delon-chart.module.ts': delonChartModuleTS,
+          'src/app/app.config.ts': appConfigTS,
           'src/app/startup.service.ts': this.genStartupService,
           'src/styles.css': ``,
           ...this.genMock
         },
-        template: 'angular-cli'
+        template: 'node'
       },
       {
         openFile: `src/app/app.component.ts`
@@ -221,12 +182,14 @@ export class CodeService {
   }
 
   openOnCodeSandbox(title: string, appComponentCode: string, includeCli: boolean = false): void {
+    appComponentCode = this.attachStandalone(appComponentCode);
     const res = this.parseCode(appComponentCode);
     const mockObj = this.genMock;
     const json = deepCopy(angularJSON);
     json.projects.demo.architect.build.options.styles.splice(0, 0, this.themePath);
     const packageJson = this.genPackage({ dependencies: [], devDependencies: [], includeCli });
-    packageJson.name = title;
+    packageJson.name = 'NG-ALAIN';
+    packageJson.description = title;
     const files: {
       [key: string]: {
         content: string;
@@ -245,44 +208,20 @@ export class CodeService {
         content: `${JSON.stringify(tsconfigJSON, null, 2)}`,
         isBinary: false
       },
-      'src/environments/environment.ts': {
-        content: environmentTS,
-        isBinary: false
-      },
       'src/index.html': {
         content: res.html,
         isBinary: false
       },
       'src/main.ts': {
-        content: includeCli ? mainCliTS : mainTS,
+        content: mainTS(res.componentName),
         isBinary: false
       },
-      'src/polyfills.ts': {
-        content: polyfillTS,
-        isBinary: false
-      },
-      'src/app/app.module.ts': {
-        content: appModuleTS(res.componentName),
-        isBinary: false
-      },
-      'src/app/global-config.module.ts': {
-        content: globalConfigTS,
+      'src/app/app.config.ts': {
+        content: appConfigTS,
         isBinary: false
       },
       'src/app/app.component.ts': {
         content: appComponentCode,
-        isBinary: false
-      },
-      'src/app/ng-zorro-antd.module.ts': {
-        content: nzZorroAntdModuleTS,
-        isBinary: false
-      },
-      'src/app/delon-abc.module.ts': {
-        content: delonABCModuleTS,
-        isBinary: false
-      },
-      'src/app/delon-chart.module.ts': {
-        content: delonChartModuleTS,
         isBinary: false
       },
       'src/app/startup.service.ts': {
@@ -300,6 +239,10 @@ export class CodeService {
       '_mock/index.ts': {
         content: mockObj['_mock/index.ts'],
         isBinary: false
+      },
+      'yarn.lock': {
+        content: yarnLock,
+        isBinary: false
       }
     };
     if (includeCli) {
@@ -307,11 +250,13 @@ export class CodeService {
         content: readme,
         isBinary: false
       };
-      files['sandbox.config.json'] = {
-        content: `${JSON.stringify(sandboxConfigJSON, null, 2)}`,
+    }
+    Object.keys(sandboxConfigJSON).forEach(key => {
+      files[key] = {
+        content: (sandboxConfigJSON as NzSafeAny)[key],
         isBinary: false
       };
-    }
+    });
     const parameters = getParameters({
       files
     });