+++ src/app/core/code/files/sandbox.ts
@@ -1,0 +1,124 @@
+export default {
+  '.codesandbox/tasks.json': `{
+    // These tasks will run in order when initializing your CodeSandbox project.
+    "setupTasks": [
+      {
+        "name": "Install Dependencies",
+        "command": "npm install"
+      }
+    ],
+
+    // These tasks can be run from CodeSandbox. Running one will open a log in the app.
+    "tasks": {
+      "start": {
+        "name": "start",
+        "command": "npm run start",
+        "runAtStart": true,
+        "preview": {
+          "port": 4200
+        }
+      },
+      "build": {
+        "name": "build",
+        "command": "ng build",
+        "runAtStart": false
+      },
+      "test": {
+        "name": "test",
+        "command": "ng test",
+        "runAtStart": false
+      },
+      "help": {
+        "name": "help",
+        "command": "ng help",
+        "runAtStart": false
+      }
+    }
+  }
+  `,
+  '.codesandbox/template.json': `{
+    "title": "Angular",
+    "description": "The quickest way to get started with Angular!",
+    "iconUrl": "https://github.com/codesandbox/sandbox-templates/blob/main/angular/.codesandbox/icon.png?raw=true",
+    "tags": [
+      "angular",
+      "javascript",
+      "typescript"
+    ],
+    "published": true
+  }`,
+  '.devcontainer/devcontainer.json': `{
+    "name": "Devcontainer",
+    "image": "ghcr.io/codesandbox/devcontainers/typescript-node:latest"
+  }
+  `,
+  '.vscode/extensions.json': `{
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=827846
+    "recommendations": ["cipchk.ng-alain-extension-pack"]
+  }
+  `,
+  '.vscode/launch.json': `{
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+      {
+        "name": "ng serve",
+        "type": "chrome",
+        "request": "launch",
+        "preLaunchTask": "npm: start",
+        "url": "http://localhost:4200/"
+      },
+      {
+        "name": "ng test",
+        "type": "chrome",
+        "request": "launch",
+        "preLaunchTask": "npm: test",
+        "url": "http://localhost:9876/debug.html"
+      }
+    ]
+  }
+  `,
+  '.vscode/tasks.json': `{
+    // For more information, visit: https://go.microsoft.com/fwlink/?LinkId=733558
+    "version": "2.0.0",
+    "tasks": [
+      {
+        "type": "npm",
+        "script": "start",
+        "isBackground": true,
+        "problemMatcher": {
+          "owner": "typescript",
+          "pattern": "$tsc",
+          "background": {
+            "activeOnStart": true,
+            "beginsPattern": {
+              "regexp": "(.*?)"
+            },
+            "endsPattern": {
+              "regexp": "bundle generation complete"
+            }
+          }
+        }
+      },
+      {
+        "type": "npm",
+        "script": "test",
+        "isBackground": true,
+        "problemMatcher": {
+          "owner": "typescript",
+          "pattern": "$tsc",
+          "background": {
+            "activeOnStart": true,
+            "beginsPattern": {
+              "regexp": "(.*?)"
+            },
+            "endsPattern": {
+              "regexp": "bundle generation complete"
+            }
+          }
+        }
+      }
+    ]
+  }
+  `
+};