+++ src/app/shared/index.ts
@@ -1,2 +1,1 @@
 export * from './components/index';
-export * from './shared.module';